Added current_user_id to cookie, which is a 8 digits random integer, used for tracking individual voters.
Extended cookie to last for 10 years, so same user coming back from the same browser can be identified.

Added a random (0-1) field in the Post, to facilitate random selection of posts.

Added votes array, which is array of hash of {current_user_id: vote) to allow statistics.

Added add_post method to Post class.

To do:
1. Add add_post control to allow remote adding of a new post. A new post should be checked against the existing ones to prevent duplicates.
2. Push to Heroku.